core: calculate fontRatio and apply it in GCText
On some low-dpi Android devices with high res (e.g. Galaxy Tab 4) the
fonts in Text-like elements appear too small with respect to the other
graphics -- also if we are using font.pointSize.

For these cases we calculate a fontRatio in ApplicationInfo that takes
dpi information into account, as proposed on
http://doc.qt.io/qt-5/scalability.html#calculating-scaling-ratio

GCText applies this factor automatically on its new fontSize property.