core, audio aac support
This patch adds the COMPRESSED_AUDIO CMake variable. By default it is set
to ogg and can be changed to aac.

When set to aac, the GCAudio item replaces .ogg by aac
and the downloadManager now get the rcc in data/voices-aac/voices-en.rcc
I uploaded aac files on gcompris.net accordingly. To make it easy, the
ogg is now available on data/voices-ogg instead of data/voices.
The data/voices is left on the server for compatibility reason.