core, add support for a full.rcc local pre downloaded resource file
If present locally the resource file 'full.rcc' is loaded (but never
downloaded). This file is present on gcompris.net/data2 as the other
resource files.

This implies that a data from an rcc can be loaded from either the
'full.rcc' or a specific one like 'words.rcc'.

Now the resource files are normalized and must include the prefix
'gcompris/data' with <qresource prefix="/gcompris/data"\>.

All the voices are now located under 'voices-ogg' instead of 'voices'
to allow us to support'voices-aac' for other platforms without having
special cases.

This patch replaces the 'isResourceRegistered' api entry by
'isDataRegistered'. We no more check a resource file is registered but
if the needed data is loaded.