balancebox: rework orientation locking
As both QTiltSensor and QRotationSensor on Android report
isFeatureSupported(AxesOrientation) == false
we honour swapped axes etc. in TiltSensor manually. To make it easy
for now we force any mobile device into landscape orientation and swap
axis if nativeOrientation is portrait.

The corresponding native functions are now moved to the
ApplicationAndroid module according to the new code structure.
structure of