balancebox, add qml-box2d as git submodule and build target
Most distributions don't provide a packaged qml-box2d 2.0.
Therefore and to simplify things we build the qml-box2d automatically
as external target and install it under lib/qml/.

Therefore also GCompris now searches also in the relative path
../lib/qml for its qml plugins on desktop.

For android we add the module correspondingly.

The git submodule for now points directly to the upstream repo on
github.