balancebox, change location of user level-file to make it readable by oher apps
The user levels file should be stored at a location where it can be
read by other applications, because users might want to share/copy
this file. Therefore we locate it under
QStandardPaths::writableLocation(QStandardPaths::GenericDataLocation)
which resolves to

     * On Android (version?): /storage/emulated/0/GCompris
     * On Linux: $HOME/local/share/GCompris

In ApplicationInfo we expose this location in getSharedWritablePath().
This location is supposed to store application-data which can be
shared with other applications (comparable to the 'My GCompris'
directory in the gtk+ version).